+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Deprecate `StateType` protocol requirement (#462) - @mjarvis
 
 **Other:**
-
+- Fix bug where automaticallySkipsRepeats is ignored when subscribing on generic `S: StoreType` (#463) - @mjarvis
 
 # 6.0.0